requirements.txt
Pillow==10.1.0
diffusers==0.23.1
transformers==4.35.2
fastapi==0.104.1
imagehash==4.3.1
numpy==1.26.2
uvicorn==0.24.0.post1
omegaconf==2.3.0
